+++ client/package.json
@@ -6,6 +6,7 @@
     "@babel/core": "7.6.0",
     "@svgr/webpack": "4.3.2",
     "@types/axios": "^0.14.0",
+    "@types/file-saver": "^2.0.1",
     "@types/jest": "24.0.18",
     "@types/js-cookie": "^2.2.2",
     "@types/node": "12.7.5",
@@ -37,6 +38,7 @@
     "eslint-plugin-react": "7.14.3",
     "eslint-plugin-react-hooks": "^1.6.1",
     "file-loader": "3.0.1",
+    "file-saver": "^2.0.2",
     "fs-extra": "7.0.1",
     "html-webpack-plugin": "4.0.0-beta.5",
     "identity-obj-proxy": "3.0.0",
@@ -71,9 +73,9 @@
     "style-loader": "1.0.0",
     "terser-webpack-plugin": "1.4.1",
     "three": "^0.107.0",
+    "ts-pnp": "1.1.4",
     "tslint": "^5.20.0",
     "tslint-react": "^4.1.0",
-    "ts-pnp": "1.1.4",
     "typescript": "3.6.3",
     "url-loader": "2.1.0",
     "webpack": "4.40.2",
